Telegram-chatbot/requirements.txt
aiogram==2.22.1
google-genai==1.54.0
python-dotenv==1.2.0
